fixed ip address detection.
Previously, the script only searched for IP addresses of adapters whose names started with 'en'. Now, it is not only much simpler, but also more general: it sets the variable in .env to use the first address returned by "hostename -i", which allows it to function properly even if you aren't connected via ethernet. Worth noting is that the first option isn't necessarily the correct option, but